move docker tests into own module
