+++ tests/test_util.py
@@ -53,51 +53,6 @@ class Utiltests(unittest.TestCase):
         """The overhead calculation returns some float greater than zero on normal execution."""
         self.assertGreater(measure_runtime_overhead(), 0)
 
-    def test_build_docker_container_timeout(self):
-        """False is returned if building a container runs into a timeout."""
-        self.assertFalse(util.build_docker_container(Path(self.problem_path, 'generator_build_timeout'),
-                                                     docker_tag='gen_bld_to',
-                                                     timeout_build=0.5,
-                                                     cache_docker_container=False))
-
-    def test_build_docker_container_failed_build(self):
-        """False is returned if building a docker container fails for any reason other than a timeout."""
-        self.assertFalse(util.build_docker_container(Path(self.problem_path, 'generator_build_error'),
-                                                     docker_tag='gen_bld_err',
-                                                     cache_docker_container=False))
-
-    def test_build_docker_container_successful_build(self):
-        """True is returned if a docker container builds successfully."""
-        self.assertTrue(util.build_docker_container(Path(self.problem_path, 'generator'),
-                                                    docker_tag='gen_succ',
-                                                    cache_docker_container=False))
-
-    def test_build_docker_container_nonexistant_path(self):
-        """False is returned if the path to the container does not exist in the file system."""
-        self.assertFalse(util.build_docker_container(Path(self.problem_path, 'foobar'),
-                                                     docker_tag='foo_bar',
-                                                     cache_docker_container=False))
-
-    def test_run_subprocess_timeout(self):
-        """run_subprocess returns None on a subprocess timeout."""
-        docker_tag = 'gen_to'
-        util.build_docker_container(Path(self.problem_path, 'generator_timeout'),
-                                    docker_tag,
-                                    cache_docker_container=False)
-        run_command = self.fight_handler.base_run_command(self.fight_handler.space_generator) + [docker_tag]
-        raw_output, _ = util.run_subprocess(run_command, input=None, timeout=2.0)
-        self.assertIsNone(raw_output)
-
-    def test_run_subprocess_execution_error(self):
-        """run_subprocess returns None if an exception is thrown during execution of the subprocess."""
-        docker_tag = 'gen_err'
-        util.build_docker_container(Path(self.problem_path, 'generator_execution_error'),
-                                    docker_tag,
-                                    cache_docker_container=False)
-        run_command = self.fight_handler.base_run_command(self.fight_handler.space_generator) + [docker_tag]
-        raw_output, _ = util.run_subprocess(run_command, input=None, timeout=10.0)
-        self.assertIsNone(raw_output)
-
     def test_update_nested_dict(self):
         """A nested dict should be updated with information from another nested dict as expected."""
         dict_to_be_expanded = {0: 1, 1: {0: 0, 1: 0}, 2: 2}
